Moved tests into their own file
